Cleaned up project
