+++ app/src/main/AndroidManifest.xml
@@ -21,9 +21,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MealListActivity" />
-        <activity android:name=".MealActivity" />
-        <activity android:name=".NewCategoryActivity" />
 
     </application>
