+++ app/src/main/java/iths/com/food/Fragments/AddMealFragment.java
@@ -151,15 +151,36 @@ public class AddMealFragment extends Fragment{
 
         long id = db.insertMeal(meal);
 
-        Toast.makeText(getActivity(), "Saved to "+meal.getCategory(), Toast.LENGTH_SHORT).show();
-        Toast.makeText(getActivity(), "Health: "+HeartRating.getHealthGrade(), Toast.LENGTH_SHORT).show();
-        Toast.makeText(getActivity(), "Taste: "+HeartRating.getTasteGrade(), Toast.LENGTH_SHORT).show();
-        Toast.makeText(getActivity(), "Name: "+name.getText().toString(), Toast.LENGTH_SHORT).show();
-        Toast.makeText(getActivity(), "Date: "+dateFormat.format(dateTime), Toast.LENGTH_SHORT).show();
-        Toast.makeText(getActivity(), "ID: "+id, Toast.LENGTH_SHORT).show();
+        db.close();
 
         //TODO: ändra så att den öppnar typ MealListFragment
         getActivity().getSupportFragmentManager().beginTransaction()
                 .replace(R.id.container, new CategoryFragment()).commit();
     }
+
+    public void updateMeal(View view) {
+
+        //VI FÅR IN ID PÅ ANNAT SÄTT
+        //String text = ((EditText) findViewById(R.id.idOfMeal)).getText().toString();
+        //id = Long.valueOf(text);
+        Long id = 3L;
+        Meal meal = db.getMeal(id);
+        meal.setHealthyScore(healthGrade);
+        meal.setTasteScore(tasteGrade);
+        meal.setName(name.getText().toString());
+        meal.setDescription(description.getText().toString());
+//PUTBACK **** meal.setCategory(spinner.getSelectedItem().toString());
+        meal.setImagePath("insert ImagePath");
+
+        // Man kan inte ändra ID eller location??
+
+        int rowsAffected = db.updateMeal(meal);
+
+        if (rowsAffected > 0) {
+            Toast.makeText(getActivity(), rowsAffected+" rows updated", Toast.LENGTH_SHORT).show();
+        }
+
+    }
 }
+
+
