+++ app/src/main/java/iths/com/food/Helper/CategoryAdapter.java
@@ -61,6 +61,8 @@ public class CategoryAdapter extends ArrayAdapter<String> {
 
         imageView.setImageResource(getContext().getResources().getIdentifier("img" + (position + 1), "drawable", getContext().getPackageName()));
 
+        db.close();
+
         return customView;
 
     }
