--- app/src/main/java/iths/com/food/MealListActivity.java
@@ -1,13 +1,0 @@
-package iths.com.food;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MealListActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_meal_list);
-    }
-}
